requirements.txt
numpy~=1.19.2
Flask==2.0.0
Pillow==8.2.0
tensorflow-cpu==2.5.0rc1
Werkzeug==2.0.0
gunicorn==20.0.4
